vllm-deploy/requirements.txt
vllm>=0.6.0
pyyaml>=6.0
requests>=2.31.0
psutil>=5.9.0